runpodgpu/requirements.txt
python-dotenv==1.0.1
requests>=2.32.3
tqdm>=4.66.5
numpy>=1.24.4
pandas>=2.2.2

torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
accelerate==0.33.0

openai-whisper==20250625
moviepy==1.0.3
yt-dlp==2025.10.22
pydub>=0.25.1
ffmpeg-python>=0.2.0

Pillow>=10.3.0
opencv-python-headless>=4.10.0.84

langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.3
langchain-chroma==0.1.4
langchain-huggingface==0.1.2
langchain-google-genai==0.3.0
langchain-text-splitters==0.3.0
google-generativeai==0.8.3

chromadb==0.5.15
sentence-transformers==3.1.1
transformers==4.45.2
scikit-learn==1.5.2
faiss-cpu==1.8.0

fpdf2==2.7.9
PyPDF2>=3.0.1
pdfminer.six>=20221105
pypdf>=4.3.1

boto3==1.35.43
botocore>=1.35.43

runpod==1.6.2

textwrap3>=0.9.2
uuid==1.30
shutilwhich>=1.1.0
typing_extensions>=4.12.2

black>=24.8.0
pytest>=8.3.3
